requirements.txt
asgiref==3.11.0
coverage==7.13.0
Django==6.0
django-filter==25.2
django-restframework==0.0.1
djangorestframework==3.16.1
factory_boy==3.3.3
Faker==39.0.0
gunicorn==23.0.0
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
pytest==9.0.2
pytest-cov==7.0.0
pytest-django==4.11.1
python-dotenv==1.2.1
sqlparse==0.5.4
tzdata==2025.3
